refactor(#4839): rename CDC connector module to MSSQL
