+++ streampipes-extensions/pom.xml
@@ -35,7 +35,7 @@
         <module>streampipes-connect-adapters-iiot</module>
 
         <module>streampipes-connectors-camel</module>
-        <module>streampipes-connectors-cdc</module>
+        <module>streampipes-connectors-mssql</module>
         <module>streampipes-connectors-filewatcher</module>
         <module>streampipes-connectors-influx</module>
         <module>streampipes-connectors-kafka</module>
