+++ streampipes-extensions/streampipes-connectors-mssql/src/main/resources/org.apache.streampipes.connect.cdc.adapter.mssql.polling/documentation.md
