+++ streampipes-extensions/streampipes-extensions-all-jvm/pom.xml
@@ -81,7 +81,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.streampipes</groupId>
-            <artifactId>streampipes-connectors-cdc</artifactId>
+            <artifactId>streampipes-connectors-mssql</artifactId>
             <version>0.99.0-SNAPSHOT</version>
         </dependency>
         <dependency>
