+++ streampipes-integration-tests/pom.xml
@@ -107,7 +107,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.streampipes</groupId>
-      <artifactId>streampipes-connectors-cdc</artifactId>
+      <artifactId>streampipes-connectors-mssql</artifactId>
       <version>0.99.0-SNAPSHOT</version>
       <scope>test</scope>
     </dependency>
